[수정] 회원가입 아이디 중복 확인
[수정] 회원가입 아이디 중복 확인

+++ frontend/chargerReservation/src/common/commonservice.ts
@@ -39,8 +39,9 @@ common.interceptors.response.use(
         const refreshToken = localStorage.getItem('refreshToken');
 
         if (!refreshToken) {
-          localStorage.clear();
-          window.location.href = '/login';
+          if (window.location.pathname !== '/register') { // 회원가입 중이 아닐 때만 이동
+              localStorage.clear();
+          }
           return Promise.reject(error);
         }
